requirements.txt
Flask==2.3.6
Flask-Session==0.5.0
python-dotenv==1.0.1
gunicorn==21.2.0
requests==2.32.0